requirements.txt
pytest
pylint
click
databricks
databricks-cli
ipython
databricks-sql-connector
databricks-connect==10.4.*
fastapi
uvicorn[standard]
pandas
sodapy
python-dotenv